Improvement

+++ reset-tag.sh
@@ -1,3 +1,8 @@
+if [[ "$1" == "" ]]; then
+	echo "Usage: reset-tag <tagname>"
+	exit
+fi
+
 git tag -d $1
 git push --delete origin $1
 git tag $1
